+++ libs/integration-interface/src/lib/topics/remote-components/v1/remote-component.model.ts
@@ -1,8 +1,8 @@
-export interface RemoteComponent { 
-    name: string;
-    baseUrl: string;
-    remoteEntryUrl: string;
-    appId: string;
-    productName: string;
-    exposedModule: string;
+export type RemoteComponent = {
+  name: string
+  baseUrl: string
+  remoteEntryUrl: string
+  appId: string
+  productName: string
+  exposedModule: string
 }